+++ src/core/layout/Layout.tsx
@@ -1,21 +1,24 @@
+import Navbar from "../components/Navbar/Navbar";
+import { AuthProvider } from "../contexts/Authcontext";
+
 interface LayoutProps {
   children: JSX.Element[] | JSX.Element
 }
 
 function Layout({children}: LayoutProps) {
   return (
-    <>
-      <header>
+    <AuthProvider>
+      <header className="flex justify-center">
+        <Navbar />
       </header>
 
-      <main className='flex'>
-
+      <main className='flex flex-col justify-center items-center'>
         <section className='w-ful'>
           {children}
         </section>
 
       </main>
-    </>
+    </AuthProvider>
   );
 }